% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{gptzero_predict_text}
\alias{gptzero_predict_text}
\title{Predict if text was generated by AI}
\usage{
gptzero_predict_text(text)
}
\arguments{
\item{text}{text to upload}
}
\value{
TODO
}
\description{
Predict if text was generated by AI
}
\examples{
\dontshow{if (GPTZeroR::has_gptzero_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Requires API Key
gptzero_predict_text() #TODO
\dontshow{\}) # examplesIf}
}
