% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{gptzero_set_key}
\alias{gptzero_set_key}
\title{Add GPTZero API key to .Renviron.}
\usage{
gptzero_set_key(key, overwrite = FALSE, install = FALSE)
}
\arguments{
\item{key}{Character. API key to add to add.}

\item{overwrite}{Defaults to FALSE. Boolean. Should existing \code{GPTZero_KEY} in Renviron be overwritten?}

\item{install}{Defaults to FALSE. Boolean. Should this be added '~/.Renviron' file?}
}
\value{
key, invisibly
}
\description{
Add GPTZero API key to .Renviron.
}
\examples{
\dontrun{
gptzero_set_key('1234')
}

}
\concept{key}
