% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPTZeroR-package.R
\docType{package}
\name{gptzeror-package}
\alias{gptzeror}
\alias{gptzeror-package}
\title{gptzeror: Identify Text Written by Large Language Models using 'GPTZero'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An R interface to the 'GPTZero' API (\url{https://gptzero.me/docs}). Allows users to classify text into human and computer written with probabilities. Formats the data into data frames where each sentence is an observation. Paragraph-level and document-level predictions are organized to align with the sentences.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/christopherkenny/gptzeror}
  \item \url{https://christophertkenny.com/gptzeror/}
  \item Report bugs at \url{https://github.com/christopherkenny/gptzeror/issues}
}

}
\author{
\strong{Maintainer}: Christopher T. Kenny \email{ctkenny@proton.me} (\href{https://orcid.org/0000-0002-9386-6860}{ORCID})

}
\keyword{internal}
