% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{key}
\alias{key}
\alias{gptzero_has_key}
\alias{gptzero_get_key}
\title{Check or Get GPTZero API Key}
\usage{
gptzero_has_key()

gptzero_get_key()
}
\value{
logical if \code{has}, key invisibly if \code{get}
}
\description{
Check or Get GPTZero API Key
}
\examples{
gptzero_has_key()
}
\concept{key}
