% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{gptzero_predict_file}
\alias{gptzero_predict_file}
\title{Predict if a file was generated by AI}
\usage{
gptzero_predict_file(file)
}
\arguments{
\item{file}{file(s) to upload. Up to 5.}
}
\value{
TODO
}
\description{
Predict if a file was generated by AI
}
\examples{
\dontshow{if (gptzeror::gptzero_has_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Requires API Key
gptzero_predict_file() #TODO
\dontshow{\}) # examplesIf}
}
