% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gptzeror-package.R
\docType{package}
\name{gptzeror-package}
\alias{gptzeror}
\alias{gptzeror-package}
\title{gptzeror: Test for Text Written by Large Language Models using GPTZero}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/christopherkenny/gptzeror}
  \item Report bugs at \url{https://github.com/christopherkenny/gptzeror/issues}
}

}
\author{
\strong{Maintainer}: Christopher T. Kenny \email{christopherkenny@fas.harvard.edu} (\href{https://orcid.org/0000-0002-9386-6860}{ORCID})

}
\keyword{internal}
